backend/requirements.txt
fastapi
uvicorn
pydantic
qdrant-client
openai
python-dotenv  
cohere
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
pinecone-client
beautifulsoup4
html2text